Test the Full Text Search queries.
They work well. Also modify the return value to List<Restaurant>. I left
the query that count the result for possibile other uses. It will be
probably removed in the next future.

Hint: the fullTextPatter of getRestaurantByString() has to be without
double quote. This means that *all* the link must be in this form
../restaurants?query=word and not ../restaurants?query="word".

Missing view integration.